backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
python-dotenv>=1.0.1
openai>=1.14.0
requests # Generally useful, included for potential future use
